pyproject.toml
[project]
name = "rwad-al-furas-job-matching-freelance-donation-system-backend"
version = "0.1.0"
description = "Rwad Al Furas Job Matching, Freelance & Donation System Backend - A comprehensive platform for job seekers, job publishers, and supporters with admin dashboard"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery[redis]>=5.5.3",
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-debug-toolbar>=5.2.0",
    "django-filter>=25.1",
    "django-redis>=6.0.0",
    "django-taggit>=6.1.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "djoser>=2.3.1",
    "drf-spectacular>=0.28.0",
    "drf-standardized-responses>=0.1.4",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.10.0",
    "python-slugify>=8.0.4",
    "whitenoise[brotli]>=6.9.0",
    "pillow>=10.4.0",
    "python-dotenv>=1.0.0",
    "redis>=5.2.1",
    "phonenumbers>=9.0.8",
    "geopy>=2.4.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
    "coverage>=7.4.0",
    "bandit>=1.7.5",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "django-extensions>=3.2.3",
    "pre-commit>=3.6.0",
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*"]

[tool.coverage.run]
source = "."
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
    "manage.py",
    "*/settings/*",
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "env",
    ".env",
    "migrations",
    "node_modules",
    "build",
    "dist",
    "*.egg-info",
    ".tox",
    ".coverage",
    "htmlcov",
    ".pytest_cache",
    ".mypy_cache",
    "staticfiles",
    "media",
]
